# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MiniBlog API
  version: 1.0.0
  description: API REST para gestionar authors, posts y comments con Express y PostgreSQL.
servers:
  - url: http://localhost:3000
    description: Local
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: API operativa
  /authors:
    get:
      summary: Listar authors
      responses:
        '200':
          description: Lista de authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
    post:
      summary: Crear author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
            example:
              name: Laura Perez
              email: laura@example.com
              bio: Editora tecnica
      responses:
        '201':
          description: Author creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
  /authors/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Obtener author por id
      responses:
        '200':
          description: Author encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
      responses:
        '200':
          description: Author actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar author
      responses:
        '204':
          description: Author eliminado
        '404':
          $ref: '#/components/responses/NotFound'
  /posts:
    get:
      summary: Listar posts
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Crear post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
            example:
              title: Testing con Vitest
              content: Contenido del post
              author_id: 1
              published: true
      responses:
        '201':
          description: Post creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
  /posts/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Obtener post por id
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '200':
          description: Post actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar post
      responses:
        '204':
          description: Post eliminado
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/author/{authorId}:
    parameters:
      - name: authorId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      summary: Listar posts por author con detalle de author
      responses:
        '200':
          description: Posts del author
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostWithAuthor'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/comments:
    get:
      summary: Listar comments
      responses:
        '200':
          description: Lista de comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentWithRelations'
    post:
      summary: Crear comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
            example:
              content: Excelente articulo
              post_id: 1
              author_id: 2
      responses:
        '201':
          description: Comment creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/comments/post/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      summary: Listar comments de un post
      responses:
        '200':
          description: Comments del post
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentWithAuthor'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Request invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: name es obligatorio
            status: 400
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Autor no encontrado
            status: 404
  schemas:
    Author:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        bio:
          type:
            - string
            - 'null'
        created_at:
          type: string
          format: date-time
    AuthorInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        bio:
          type: string
    Post:
      type: object
      properties:
        id:
          type: integer
        author_id:
          type: integer
        title:
          type: string
        content:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
    PostInput:
      type: object
      required:
        - title
        - content
        - author_id
      properties:
        title:
          type: string
        content:
          type: string
        author_id:
          type: integer
        published:
          type: boolean
    PostWithAuthor:
      allOf:
        - $ref: '#/components/schemas/Post'
        - type: object
          properties:
            author:
              $ref: '#/components/schemas/Author'
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        post_id:
          type: integer
        author_id:
          type:
            - integer
            - 'null'
        created_at:
          type: string
          format: date-time
    CommentInput:
      type: object
      required:
        - content
        - post_id
        - author_id
      properties:
        content:
          type: string
        post_id:
          type: integer
        author_id:
          type: integer
    CommentWithAuthor:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            author:
              type:
                - object
                - 'null'
    CommentWithRelations:
      allOf:
        - $ref: '#/components/schemas/CommentWithAuthor'
        - type: object
          properties:
            post:
              type: object
    Error:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer